FIX: More accurate matching for tag names (#2)
Previously, if you had the following tag -> icon mappings in the setting:

```
education-team,users
us,globe
```

The `us` tag would not work, because the returned match would be the
first row. This PR fixes it so matching only checks the tag names (i.e.
only the first column in the CSV-like data).